backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-MySQLdb==2.0.0
Flask-Session==0.8.0
fonttools==4.57.0
fsspec==2025.3.2
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.8
lxml==5.3.2
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
msgspec==0.19.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.1
oauthlib==3.2.2
openai==1.72.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
Pillow==10.1.0
proglog==0.1.11
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.8.0
PyMuPDF==1.25.4
PyMySQL==1.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.1.0
uuid==1.30
websocket-client==1.2.1
Werkzeug==3.1.3
XlsxWriter==3.2.2
zhipuai==2.1.5.20250410
zope.event==5.0
zope.interface==7.2
